pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ankiflow"
version = "0.1.0"
description = "Facilitador de tradução e integração automática com o Anki"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "Kevin Pontes" }
]
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.7",
    "rich>=13.7.0",
]

[project.scripts]
ankiflow = "src.main:cli"

[tool.setuptools]
packages = ["src", "src.config"]